model and class diagram edited. images and price moved from vehicle into offer
